Use 'int' for indices
No need more than 10^6 values, the full unsigned 64-bit range is overkill.
For marker values, instead of std::numeric_limits<size_t>::max(), use -1.

+++ src/apps/dashcastx/pipeliner_dashcastx.cpp
@@ -142,7 +142,7 @@ std::unique_ptr<Pipeline> buildPipeline(const IConfig &config) {
 	}
 
 	int numDashInputs = 0;
-	for (size_t i = 0; i < demux->getNumOutputs(); ++i) {
+	for (int i = 0; i < demux->getNumOutputs(); ++i) {
 		auto const metadataDemux = safe_cast<const MetadataPktLibav>(demux->getOutput(i)->getMetadata());
 		if (!metadataDemux) {
 			Log::msg(Warning, "[%s] Unknown metadataDemux for stream %s. Ignoring.", g_appName, i);
