+++ src/apps/mp42tsx/pipeliner_mp42ts.cpp
@@ -30,7 +30,7 @@ void declarePipeline(Pipeline &pipeline, const mp42tsXOptions &opt) {
 	auto demux = pipeline.addModule<Demux::LibavDemux>(opt.url);
 	auto m2tsmux = pipeline.addModule<Mux::GPACMuxMPEG2TS>(GF_FALSE, 0);
 	auto sink = createSink(isHLS);
-	for (size_t i = 0; i < demux->getNumOutputs(); ++i) {
+	for (int i = 0; i < demux->getNumOutputs(); ++i) {
 		if (demux->getOutput(i)->getMetadata()->isVideo()) { //FIXME
 			pipeline.connect(demux, i, m2tsmux, 0);
 			break; //FIXME
